refactor: cookie 및 hooks,util 디렉토리 lib 폴더 하위로 이동
