+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
 import store from "./store/index";
 import App from "./App";
-import { axeTester } from "./util/axeTester";
+import { axeTester } from "./lib/util/axeTester";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
